create many-to-many betw tables(product&order), add an exception handle for creating an order

+++ store/src/main/java/com/kasperovich/store/repository/OrderRepository.java
@@ -2,10 +2,11 @@ package com.kasperovich.store.repository;
 
 import com.kasperovich.store.model.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
+import org.springframework.transaction.annotation.Transactional;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
-
-
-
+public interface OrderRepository extends JpaRepository<Order, Long> {
 
 }
